backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.9

# Validation & settings
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Task queue
celery==5.4.0
redis==5.1.1
flower==2.0.1

# HTTP client (for external APIs)
httpx==0.27.2

# Dev/testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2
coverage==7.6.1
aiosqlite==0.20.0
bcrypt==4.0.1

pandas==2.2.3
pdfplumber==0.11.4
google-generativeai==0.8.3
google-genai
python-multipart==0.0.12   # needed for FastAPI file uploads
celery[redis]==5.4.0        # update existing celery line
weasyprint==62.3
pydyf==0.10.0
sendgrid==6.11.0
boto3==1.35.36
jinja2==3.1.4
stripe==10.12.0
websockets==13.1